+++ Sources/SwiftGenCLI/Config/Config+Run.swift
@@ -10,10 +10,10 @@ import StencilSwiftKit
 import SwiftGenKit
 
 extension Config {
-  public func runCommands(verbose: Bool, logger: (LogLevel, String) -> Void = logMessage) throws {
+  public func runCommands(logLevel: CommandLogLevel, logger: (LogLevel, String) -> Void = logMessage) throws {
     let errors = commands.parallelCompactMap { cmd, entry -> Swift.Error? in
       do {
-        try run(parserCommand: cmd, entry: entry, verbose: verbose, logger: logger)
+        try run(parserCommand: cmd, entry: entry, logLevel: logLevel, logger: logger)
         return nil
       } catch {
         return error
@@ -26,17 +26,17 @@ extension Config {
       throw Error.multipleErrors(errors)
     }
   }
-
+  
   private func run(
     parserCommand: ParserCLI,
     entry: ConfigEntry,
-    verbose: Bool,
+    logLevel: CommandLogLevel,
     logger: (LogLevel, String) -> Void
   ) throws {
     var entry = entry
 
     entry.makeRelativeTo(inputDir: inputDir, outputDir: outputDir)
-    if verbose {
+    if logLevel == .verbose {
       for item in entry.commandLine(forCommand: parserCommand.name) {
         logger(.info, " $ \(item)")
       }